+++ examples/skills/game/src/index.ts
@@ -1,8 +1,7 @@
-import {Bot} from '../../../src';
-import skillGameConfig from '../../config/skillGameConfig';
-import skillGameParam from '../../config/skillGameParam';
+import {Bot} from 'umbot';
+import skillGameConfig from './config/skillGameConfig';
+import skillGameParam from './config/skillGameParam';
 import {GameController} from './controller/GameController';
-import {IncomingMessage, ServerResponse} from 'http';
 
 const bot = new Bot();
 bot.initConfig(skillGameConfig());
@@ -16,6 +15,4 @@ bot.initBotController(logic);
 //     isShowTime: true,
 // }
 // bot.test(params);
-module.exports = async (req: IncomingMessage, res: ServerResponse) => {
-    bot.start(req, res)
-};
+bot.start('localhost', 3000);