+++ src/build.ts
@@ -1,8 +1,8 @@
-import {BotTest, IBotTestParams} from "./core/BotTest";
-import {IAppConfig, IAppParam, mmApp} from "./mmApp";
-import {BotController} from "./controller";
-import {Bot} from "./core";
-import {IncomingMessage, ServerResponse} from "http";
+import { BotTest, IBotTestParams } from "./core/BotTest";
+import { IAppConfig, IAppParam, mmApp } from "./mmApp";
+import { BotController } from "./controller";
+import { Bot } from "./core";
+import { IncomingMessage, ServerResponse } from "http";
 
 /**
  * Набор методов, упрощающих запуск приложения
@@ -14,8 +14,10 @@ import {IncomingMessage, ServerResponse} from "http";
  * - dev Запуск в режиме тестирования, с использованием консоли, в качестве тестового окружения.
  * - dev-online Запуск в режиме тестирования, с использованием webhook, в качестве тестового окружения.
  * - prod Запуск в релизном режиме
+ * - dev-online-old Запуск в режиме тестирования, с использованием webhook, в качестве тестового окружения в старом режиме.
+ * - prod-old Запуск в релизном режиме в старом режиме
  */
-export type TMode = 'dev' | 'dev-online' | 'prod';
+export type TMode = 'dev' | 'dev-online' | 'prod' | 'dev-online-old' | 'prod-old';
 
 /**
  * Настройки приложения
@@ -50,7 +52,7 @@ function _initParam(bot: Bot | BotTest, config: IConfig): void {
  * @param {IConfig} config Конфигурация приложения
  * @param {TMode} mode Режим работы приложения
  */
-export function run(config: IConfig, mode: TMode = 'prod'): unknown {
+export function run(config: IConfig, mode: TMode = 'prod', hostname: string = 'localhost', port: number = 3000): unknown {
     let bot: BotTest | Bot;
     switch (mode) {
         case "dev":
@@ -63,16 +65,27 @@ export function run(config: IConfig, mode: TMode = 'prod'): unknown {
             bot.initTypeInGet();
             _initParam(bot, config);
             mmApp.setDevMode(true);
-            module.exports = async (req: IncomingMessage, res: ServerResponse) => {
-                bot.start(req, res)
-            };
-            return module;
+            return bot.start(hostname, port);
         case "prod":
             bot = new Bot();
             bot.initTypeInGet();
             _initParam(bot, config);
+            return bot.start(hostname, port);
+        case "dev-online-old":
+            bot = new Bot();
+            bot.initTypeInGet();
+            _initParam(bot, config);
+            mmApp.setDevMode(true);
             module.exports = async (req: IncomingMessage, res: ServerResponse) => {
-                bot.start(req, res)
+                bot.startOld(req, res)
+            };
+            return module;
+        case "prod-old":
+            bot = new Bot();
+            bot.initTypeInGet();
+            _initParam(bot, config);
+            module.exports = async (req: IncomingMessage, res: ServerResponse) => {
+                bot.startOld(req, res)
             };
             return module;
     }